Added __vfprintf_total_size, fixed vsnprintf return value
